V1.6.0 Added the following
Large and small numbers are now displayed/entered using exponents. The '.' button becomes a 'E' button once a decimal is entered so you can enter an exponent.

As soon as an operation is entered (+, -, x, ÷ and CA), the screen autoreturns to panel 1 so you cam right away start entering number without having to swipe.